[conpliance] While novel suggestion to test invalid usage of private classes of the jdk, this needs removed
If this is to be tested, it should be done so in a mock, integration test where jdk is downgraded to earlier or some other means.  Besides, this overlaps with sonar as it is well known not to use the internal sun.* classes as it is not portable.  While this just ignores the test entirely by removal, the point is to show we can build on jdk 11 and above.